Refactor get_children to increase speed
The get_children method was slowing down the entire process of building the tree.

The main catch was iterating each list element to find which ones have certain parent.

The time complexity of this method is quadratic time O(n^2). For big tree (over 9K elements) it resulted in over 80M operations which is slow.

Searching entire array multiple time is not necessary, we could build a library of parent-child and use direct access to its elements to pick children. Time complexity is reduced to linear time.